dashboard/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
networkx>=3.2
pandas>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
numpy>=1.24.0
